requirements.txt
# Kern
gym==0.17.3
JPype1==1.2.1
pyjnius==1.6.1

# PyTorch / RL-Stack
# Wenn du bereits ein funktionierendes GPU/ROCm-Torch hast, lass 'torch*' lieber weg.
#stable-baselines3==1.0
#torch==2.4.1      # <- NUR setzen, wenn du bewusst über pip installieren willst
#torchvision==0.19.1
#torchaudio==2.4.1

# Wissenschaftsstack
numpy==1.22.4
pandas==1.2.3
matplotlib==3.4.1
seaborn==0.11.1
tqdm==4.60.0

# Logging/Training optional
tensorboard==2.4.1
wandb==0.10.24
cleanrl==0.4.3

# Sonstiges
pyglet==1.5.0
Pillow==8.2.0
